+++ docs/changes.md
@@ -1,4 +1,7 @@
 # 最新动态
+* 2019/01/30
+  * 增加binding context
+
 * 2019/01/29
   * 增加navigator